requirements.txt
Flask==2.0.3
Flask_Cors==3.0.10
flask_restx==0.5.1
Flask_SQLAlchemy==2.5.1
jenkinsapi==0.3.11
PyYAML==6.0
SQLAlchemy==1.4.39
